Barebones app window + gui based on stripped down Nanoforge
Using a stripped down version of the nanoforge rewrite as a starting point for this app. It has some code that's overkill for an asm_pc editor, but that's fine. Doing this saves a ton of time vs creating the app from scratch. Long term I hope to deprecate this app anyway and move it into Nanoforge. Not doing that initially because I don't want to deal with the extra complication of integrating it with projects / edit tracking / etc quite yet. See readme for more info.